+++ go_to_office/lib/model/seat.dart
