+++ go_to_office/lib/screens/admin/seat.dart
@@ -1,17 +1,9 @@
 import 'package:flutter/material.dart';
 
-void main(seatId, officeId) {
-  runApp(new MaterialApp(
-    title: "Office Manager",
-    home: new SeatPage(seatId, officeId),
-  ));
-}
-
 class SeatPage extends StatefulWidget {
-
-  SeatPage(this.seatId, this.officeId );
-    final String seatId;
-    final String officeId;
+  SeatPage(this.seatId, this.officeId);
+  final String seatId;
+  final String officeId;
 
   @override
   State<StatefulWidget> createState() => _SeatPageState(seatId, officeId);
@@ -20,17 +12,18 @@ class SeatPage extends StatefulWidget {
 class _SeatPageState extends State<SeatPage> {
   final String seatId;
   final String officeId;
-  String roomId ;
-  String seatLocation ;
-   _SeatPageState(this.seatId, this.officeId);
+  String roomId;
+  String seatLocation;
+  _SeatPageState(this.seatId, this.officeId);
   @override
   Widget build(BuildContext context) {
     this.roomId = '1';
-    if (this.seatId != null) {  // call API to get get office also get the office name
-    this.seatLocation = 'Near a window';
-  } else {
-    this.seatLocation = '';
-  }
+    if (this.seatId != null) {
+      // call API to get get office also get the office name
+      this.seatLocation = 'Near a window';
+    } else {
+      this.seatLocation = '';
+    }
 
     // Object officateAssreddress = {};
     return Scaffold(
@@ -42,8 +35,8 @@ class _SeatPageState extends State<SeatPage> {
             padding: EdgeInsets.all(15.0),
             child: Column(
               children: <Widget>[
-                  TextFormField(
-                  initialValue: this.seatLocation,
+                TextFormField(
+                    initialValue: this.seatLocation,
                     decoration: InputDecoration(
                         hintText: 'Please insert Seat Loaction in the room'),
                     onChanged: (String inputString) {
